crud completo de productos

+++ app_productos/models.py
@@ -9,7 +9,8 @@ from django.core.validators import MinValueValidator
 class Producto(models.Model):
     nombre = models.CharField(max_length=40)
     descripcion = models.CharField(max_length=150)
-    precio_unitario = models.FloatField(validators=[MinValueValidator(0.0)])
+    precio_unitario = models.IntegerField(validators=[MinValueValidator(0)])
 
     def __unicode__(self):
         return '{}'.format(self.nombre)
+
